+++ lib/astarte_device_sdk/bson/serializer.c
@@ -4,14 +4,14 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-#include "bson_serializer.h"
+#include "bson/serializer.h"
 
 #include <stdio.h>
 #include <stdlib.h>
 
 #include <zephyr/sys/byteorder.h>
 
-#include "bson_types.h"
+#include "bson/types.h"
 #include "log.h"
 
 ASTARTE_LOG_MODULE_REGISTER(bson_serializer, CONFIG_ASTARTE_DEVICE_SDK_BSON_LOG_LEVEL);