+++ lib/astarte_device_sdk/data_deserialize.c
@@ -9,8 +9,8 @@
 #include <stdlib.h>
 #include <string.h>
 
-#include "bson_deserializer.h"
-#include "bson_types.h"
+#include "bson/deserializer.h"
+#include "bson/types.h"
 #include "interface_private.h"
 #include "mapping_private.h"