+++ lib/astarte_device_sdk/device_rx.c
@@ -9,7 +9,7 @@
 #include <zlib.h>
 #endif
 
-#include "bson_deserializer.h"
+#include "bson/deserializer.h"
 #include "data_validation.h"
 #ifdef CONFIG_ASTARTE_DEVICE_SDK_PERMANENT_STORAGE
 #include "storage/prop.h"